+++ models/Recipe.js
@@ -16,14 +16,11 @@ Recipe.init(
       allowNull: false
     },
     description: {
-      type: DataTypes.OBJECT
+      type: DataTypes.STRING,
     },
     ingredients: {
-      type: DataTypes.OBJECT,
-      allowNull: false
-    },
-    instructions: {
-      type: DataTypes.OBJECT,
+      type: DataTypes.STRING,
+
       allowNull: false
     },
     image: {
