+++ package.json
@@ -5,9 +5,8 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-
     "start": "node server.js",
-    "seed": "node seeds/seed.js"
+    "seed": "node ./seeds/seeds.js"
   },
   "repository": {
     "type": "git",
@@ -19,7 +18,6 @@
   "bugs": {
     "url": "https://github.com/jacih/Project-Two/issues"
   },
-
   "dependencies": {
     "bcrypt": "^5.0.0",
     "connect-session-sequelize": "^7.0.4",
@@ -32,6 +30,6 @@
     "sequelize": "^6.3.5",
     "recipe-scraper": "^2.1.1"
   },
-  "homepage": "https://github.com/jacih/Project-Two#readme"
-
+  "homepage": "https://github.com/jacih/Project-Two#readme",
+  "devDependencies": {}
 }
